requirementsDock.txt
requests>=2.18.4
gtts>=2.0.1
urbandict>=0.5
gsearch>=1.6.0
wikipedia>=1.4.0
speedtest-cli>=2.0.2
zalgo_text>=0.6
hastebin.py>=0.2
pybase64>=0.4.0
google_images_download>=2.4.2
hachoir
Pillow>=5.3.0
telegraph
sqlalchemy>=1.2
gTTS>=2.0.1
googletrans>=2.4.0
spongemock
gTTS-token>=1.1.3
psycopg2
python-dotenv
aiohttp
telethon-session-sqlalchemy>=0.2.6
cryptg
pyfiglet
cowsay
google-api-python-client
google-auth-oauthlib
lmgtfy
incremental
emoji
language_check
pytube
dnspython
cowpy
